refactor: refactor base simulations
