+++ x/ecocredit/base/simulation/msg_update_class_fee.go
@@ -1,0 +1,86 @@
+package simulation
+
+import (
+	"math/rand"
+
+	"github.com/cosmos/cosmos-sdk/baseapp"
+	codectypes "github.com/cosmos/cosmos-sdk/codec/types"
+	simappparams "github.com/cosmos/cosmos-sdk/simapp/params"
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	simtypes "github.com/cosmos/cosmos-sdk/types/simulation"
+	govtypes "github.com/cosmos/cosmos-sdk/x/gov/types/v1"
+	"github.com/cosmos/cosmos-sdk/x/simulation"
+	"github.com/regen-network/regen-ledger/x/ecocredit"
+	types "github.com/regen-network/regen-ledger/x/ecocredit/base/types/v1"
+	"github.com/regen-network/regen-ledger/x/ecocredit/simulation/utils"
+)
+
+const OpWeightMsgUpdateClassFee = "op_weight_msg_update_class_fee" //nolint:gosec
+
+var TypeMsgUpdateClassFee = sdk.MsgTypeURL(&types.MsgUpdateClassFee{})
+
+const WeightUpdateClassFee = 33
+
+// SimulateMsgUpdateClassFee generates a MsgToggleClassAllowlist with random values.
+func SimulateMsgUpdateClassFee(ak ecocredit.AccountKeeper, bk ecocredit.BankKeeper, govk ecocredit.GovKeeper,
+	qryClient types.QueryServer, authority sdk.AccAddress) simtypes.Operation {
+	return func(
+		r *rand.Rand, app *baseapp.BaseApp, sdkCtx sdk.Context, accs []simtypes.Account, chainID string,
+	) (simtypes.OperationMsg, []simtypes.FutureOperation, error) {
+		proposer, _ := simtypes.RandomAcc(r, accs)
+		proposerAddr := proposer.Address.String()
+
+		spendable, account, op, err := utils.GetAccountAndSpendableCoins(sdkCtx, bk, accs, proposerAddr, TypeMsgUpdateClassFee)
+		if spendable == nil {
+			return op, nil, err
+		}
+
+		params := govk.GetDepositParams(sdkCtx)
+		deposit, skip, err := utils.RandomDeposit(r, sdkCtx, ak, bk, params, proposer.Address)
+		switch {
+		case skip:
+			return simtypes.NoOpMsg(ecocredit.ModuleName, TypeMsgUpdateClassFee, "skip deposit"), nil, nil
+		case err != nil:
+			return simtypes.NoOpMsg(ecocredit.ModuleName, TypeMsgUpdateClassFee, "unable to generate deposit"), nil, err
+		}
+
+		fee := utils.RandomFee(r)
+		if fee.Amount.IsZero() {
+			return simtypes.NoOpMsg(ecocredit.ModuleName, TypeMsgUpdateClassFee, "invalid proposal message"), nil, err
+		}
+
+		proposalMsg := types.MsgUpdateClassFee{
+			Authority: authority.String(),
+			Fee:       &fee,
+		}
+
+		any, err := codectypes.NewAnyWithValue(&proposalMsg)
+		if err != nil {
+			return simtypes.NoOpMsg(ecocredit.ModuleName, TypeMsgUpdateClassFee, err.Error()), nil, err
+		}
+
+		msg := &govtypes.MsgSubmitProposal{
+			Messages:       []*codectypes.Any{any},
+			InitialDeposit: deposit,
+			Proposer:       proposerAddr,
+			Metadata:       simtypes.RandStringOfLength(r, 10),
+		}
+
+		txCtx := simulation.OperationInput{
+			R:               r,
+			App:             app,
+			TxGen:           simappparams.MakeTestEncodingConfig().TxConfig,
+			Cdc:             nil,
+			Msg:             msg,
+			MsgType:         msg.Type(),
+			Context:         sdkCtx,
+			SimAccount:      *account,
+			AccountKeeper:   ak,
+			Bankkeeper:      bk,
+			ModuleName:      ecocredit.ModuleName,
+			CoinsSpentInMsg: spendable,
+		}
+
+		return utils.GenAndDeliverTxWithRandFees(r, txCtx)
+	}
+}
